Upload diff as artifact
Compares between the latest uploaded diff and the current diff, instead of comparing commits.
Comparing commits is problematic because when we force push, the previous commit no longer exists and the diff is identified to be the whole PR instead of only the difference between the last push and the previous one. Comparing the last uploaded diff with the current diff addresses that.